Updated remove.py

+++ remove.py
@@ -12,7 +12,7 @@ def arg_parse():
     '''
 
     parser = argparse.ArgumentParser(
-        description="Generates the synopsis of folder contents"
+        description="Generates the synopsis of folder contents. Either one of formats and ref arguments is mandatory"
     )
 
     parser.add_argument('-i', 
@@ -21,16 +21,15 @@ def arg_parse():
                         help="Full path of input directory to summarize")
 
     parser.add_argument('-formats',
-                        required=True,
                         type=str,
                         default="",
                         help="Enter the file formats to be removed from the directory structure")
 
-    parser.add_argument('-refs',
+    parser.add_argument('-ref',
                         choices=['y','n'],
                         type=str,
-                        default='n',
-                        help="full path of output directory")     
+                        default="",
+                        help="Would you like to remove empty directories? Enter y/n")     
 
     parsed_args = parser.parse_args()
     return parsed_args
@@ -43,37 +42,39 @@ def main():
     desktop_logs_dir = make_desktop_logs_dir()
     log_name_source = os.path.join(desktop_logs_dir, log_name_source_)
 
-    if not os.path.isdir(input_path):
-        print(' - Input must be a directory/folder - exiting!')
-        generate_log(log_name_source, ' - Input must be a directory/folder - exiting!')
-        sys.exit()
-    
-    remove_bad_files(input_path, log_name_source)
+    if args.formats or args.ref:
+        if not os.path.isdir(input_path):
+            print(' - Input must be a directory/folder - exiting! - exiting remove.py!')
+            generate_log(log_name_source, ' - Input must be a directory/folder - exiting remove.py!')
+            sys.exit()
+        
+        remove_bad_files(input_path, log_name_source)
+        
+        if args.formats:
+            formats = list(str(args.formats).strip().split(' '))    
+            for root, subdirs, files in os.walk(input_path):
+                
+                if files == () or files == []:
+                    continue
 
-    formats = list(str(args.formats).strip().split(' '))
-    for root, subdirs, files in os.walk(input_path):
-        try:
-            
-            if files == () or files == []:
-                continue
-
-            for file in files:
-                if str(os.path.splitext(file)[1]).lower() in formats:
-                    print(f"File {file} removed from {root}")
-                    generate_log(log_name_source, f"File {file} removed from {root}")
-                    os.remove(os.path.join(root,file))
-            
-            
-            if args.refs == 'y':
+                for file in files:
+                    if str(os.path.splitext(file)[1]).lower() in formats:
+                        print(f"File {file} removed from {root}")
+                        generate_log(log_name_source, f"File {file} removed from {root}")
+                        os.remove(os.path.join(root,file))
+        
+        if args.ref == 'y':
+            for root, subdirs, files in os.walk(input_path):    
                 for subdir in subdirs:
                     l = len(os.listdir(os.path.join(root,subdir)))
                     if l == 0:
                         print(f"Empty Folder {subdir} removed from {root}")
                         generate_log(log_name_source, f"Empty Folder {subdir} removed from {root}")
                         os.rmdir(os.path.join(root,subdir))
-        
-        except:
-            continue
+    else:
+        print(' - Either "formats" or "refs" argument is mandatory! - exiting remove.py!')
+        generate_log(log_name_source, ' - Either "formats" or "refs" argument is mandatory! - exiting remove.py!')
+        sys.exit()
 
 if __name__ == "__main__":
     main()
